Attempt at bringing back old compile method

+++ src/zquest.cpp
@@ -448,7 +448,7 @@ int32_t alignment_arrow_timer=0;
 int32_t  Flip=0,Combo=0,CSet=2,First[3]= {0,0,0},current_combolist=0,current_comboalist=0,current_mappage=0;
 int32_t  Flags=0,Flag=1,menutype=(m_block);
 int32_t MouseScroll = 0, SavePaths = 0, CycleOn = 0, ShowGrid = 0, GridColor = 0, TileProtection = 0, InvalidStatic = 0, NoScreenPreview = 0, MMapCursorStyle = 0, BlinkSpeed = 20, UseSmall = 0, RulesetDialog = 0, EnableTooltips = 0, 
-	ShowFFScripts = 0, ShowSquares = 0, ShowInfo = 0, skipLayerWarning = 0, WarnOnInitChanged = 0, DisableLPalShortcuts = 0;
+	ShowFFScripts = 0, ShowSquares = 0, ShowInfo = 0, skipLayerWarning = 0, WarnOnInitChanged = 0, DisableLPalShortcuts = 0, EnableCompileConsole = 0;
 int32_t FlashWarpSquare = -1, FlashWarpClk = 0; // flash the destination warp return when ShowSquares is active
 uint8_t ViewLayer3BG = 0, ViewLayer2BG = 0; 
 bool Vsync = false, ShowFPS = false;
@@ -24689,10 +24689,17 @@ int32_t onCompileScript()
 				break;
 			}
 			parser_console.kill();
-			parser_console.Create("ZScript Parser Output", 600, 200, NULL, "ZConsole.exe");
-			parser_console.cls(CConsoleLoggerEx::COLOR_BACKGROUND_BLACK);
-			parser_console.gotoxy(0,0);
-			zconsole_info("External ZScript Parser\n");
+			if (EnableCompileConsole) 
+			{
+				parser_console.Create("ZScript Parser Output", 600, 200, NULL, "ZConsole.exe");
+				parser_console.cls(CConsoleLoggerEx::COLOR_BACKGROUND_BLACK);
+				parser_console.gotoxy(0,0);
+				zconsole_info("External ZScript Parser\n");
+			}
+			else
+			{
+				box_start(1, "Compile Progress", lfont, sfont,true);
+			}
 			clock_t start_compile_time = clock();
 			char const* noclose = "-noclose";
 			char const* argv[5] = {"-input", "tmp", "-linked", NULL, NULL};
@@ -24747,12 +24754,19 @@ int32_t onCompileScript()
 				"Compile took %s seconds (%ld cycles)%s",
 				code, code ? "failure" : "success",
 				tmp, end_compile_time - start_compile_time,
-				code ? "\nCompilation failed. See console for details." : "");
+				code ? (EnableCompileConsole?"\nCompilation failed. See console for details.":"\nCompilation failed.") : "");
 			
 			if(!code)
 			{
 				read_compile_data(stypes, scripts);
-				parser_console.kill();
+				if (EnableCompileConsole) 
+				{
+					parser_console.kill();
+				}
+			}
+			if (!EnableCompileConsole)
+			{
+				box_end(true);
 			}
 			compile_sfx(!code);
 			
@@ -29983,6 +29997,7 @@ int32_t main(int32_t argc,char **argv)
 	chop_path(tmusicpath);
 	
 	DisableLPalShortcuts        = zc_get_config("zquest","dis_lpal_shortcut",0);
+	EnableCompileConsole        = zc_get_config("zquest","extern_compile_console",0);
 	MouseScroll					= zc_get_config("zquest","mouse_scroll",0);
 	WarnOnInitChanged			  = zc_get_config("zquest","warn_initscript_changes",1);
 	InvalidStatic				  = zc_get_config("zquest","invalid_static",1);
@@ -32252,6 +32267,7 @@ int32_t save_config_file()
     set_config_string("zquest","last_timed_save",last_timed_save);
     set_config_int("zquest","mouse_scroll",MouseScroll);
 	set_config_int("zquest","dis_lpal_shortcut",DisableLPalShortcuts);
+	set_config_int("zquest","extern_compile_console",EnableCompileConsole);
     set_config_int("zquest","warn_initscript_changes",WarnOnInitChanged);
     set_config_int("zquest","invalid_static",InvalidStatic);
 //	set_config_int("zquest","cursorblink_style",MMapCursorStyle); // You cannot do this unless the value is changed by the user via the GUI! -Z
